+++ CHANGES.rst
@@ -17,6 +17,8 @@ Unreleased
 * Fix #130: explicitly request fields from Facebook. 
 * Adjusted naming of default scope for :class:`.oauth2.Facebook` to Facebook v2 API
 * Added support for :attr:`.User.email` to the :class:`.oauth1.Bitbucket` provider.
+* Removed support for :attr:`.User.birth_date` and :attr:`.User.gender`
+  from the :class:`.oauth1.Yahoo` provider.
 
 Version 0.1.0
 -------------